requirements.txt
Faker==36.1.0
pytest==8.3.4
pytest-rerunfailures==15.0
requests==2.32.3
selenium==4.31.0
webdriver-manager==4.0.2
allure-pytest==2.13.5
pytest-regressions==2.7.0
pillow==11.2.1
numpy==2.2.4
pydantic[email]
jsonschema==4.23.0